apiwrap::cstringを公開クラスにする
