+++ sakura_core/apiwrap/cstring.cpp
@@ -1,0 +1,41 @@
+﻿/*! @file */
+/*
+	Copyright (C) 2023, Sakura Editor Organization
+
+	This software is provided 'as-is', without any express or implied
+	warranty. In no event will the authors be held liable for any damages
+	arising from the use of this software.
+
+	Permission is granted to anyone to use this software for any purpose,
+	including commercial applications, and to alter it and redistribute it
+	freely, subject to the following restrictions:
+
+		1. The origin of this software must not be misrepresented;
+		   you must not claim that you wrote the original software.
+		   If you use this software in a product, an acknowledgment
+		   in the product documentation would be appreciated but is
+		   not required.
+
+		2. Altered source versions must be plainly marked as such,
+		   and must not be misrepresented as being the original software.
+
+		3. This notice may not be removed or altered from any source
+		   distribution.
+ */
+#include "StdAfx.h"
+#include "apiwrap/cstring.hpp"
+
+namespace apiwrap {
+
+LPCWSTR get_psz_or_null(const cstring& str) noexcept
+{
+	if (const auto text = static_cast<std::wstring_view>(str);
+		0 < text.length())
+	{
+		return text.data();
+	}
+
+	return nullptr;
+}
+
+} //end of namespace apiwrap
